Split backend of sdks to module and keeper.
