+++ sdk/ownership/sdk.go
@@ -21,7 +21,7 @@ func New(client *cosmos.Client) *SDK {
 // List returns all ownerships.
 func (s *SDK) List() ([]*ownership.Ownership, error) {
 	var ownerships []*ownership.Ownership
-	if err := s.client.Query("custom/"+backendName+"/list", nil, &ownerships); err != nil {
+	if err := s.client.Query("custom/"+ModuleName+"/list", nil, &ownerships); err != nil {
 		return nil, err
 	}
 	return ownerships, nil
